Tweaking the tests
Changed the tests to be one big module so that I can add setUp and
tearDown, if desired.  My intention was to use node-ftpd and integrate
it into the tests, but it seems that the SIZE support has not been
generated (yet).

Also, seems that there is a bug in pget for large files - md5 doesn't
match.  Going to have to figure out exactly why.